Bug fixed in mesa.py, kip_cont function. Also, the functionality of this tool and docstring have been updated. SJ
git-svn-id: svn://forum.astro.keele.ac.uk/utils/pylib@2069 2e4cce12-393c-0410-8681-c83ea65bbeac

+++ mesa.py
@@ -291,7 +291,9 @@ class star_log(DataPlot):
         slname  = self.slname
         slaname = slname+'sa'
         if self.clean_starlog and os.path.exists(sldir+'/'+slaname):
-            os.remove(sldir+'/'+slaname)
+            jonesmod=str(raw_input("Clean star.log to make new star.logsa? (y/n)"))
+            if jonesmod == 'y':
+                os.remove(sldir+'/'+slaname)
             
         if not os.path.exists(sldir+'/'+slaname):
             print 'No star.logsa file found, create new one from star.log.'
@@ -607,16 +609,47 @@ class star_log(DataPlot):
 		elif xax == 'model':
 		    pyl.xlabel('model number')
 
-    def kip_cont(self,xlims=[0.,0.],ylims=[0.,0.],xres=50,yres=2000,ixaxis='log_time_left',mix_zones=5,burn_zones=50):
+    def kip_cont(self,modstart,modstop,outfile,xlims=[0.,0.],ylims=[0.,0.],xres=50,yres=2000,ixaxis='log_time_left',mix_zones=5,burn_zones=50):
         '''This function creates a Kippenhahn plot with energy flux using
-        contours. For this version you need to have printed the columns which
-        have the form burn_qtop_n and burn_type_n, where n is an integer.
-        If you have only epsnuc_M_n columns there's a script for that too.
-        If you have only conv_mx1_bot, then either re-run or let me know'''
+        contours.
+        For a more comprehensive plot, your star.log file should contain columns
+        called "mix_type_n","mix_qtop_n","burn_type_n" and "burn_qtop_n".
+        The number of columns (i.e. the bbiggest value of n) is what goes in the
+        arguments as mix_zones and burn_zones.
+        DO NOT WORRY! if you do not have these columns, just leave the default
+        values alone and the script should recognise that you do not have these
+        columns and make the most detailed plot that is available to you.
 
-        xxyy=[self.get('star_age'),self.get('star_age')]
+        modstart:        model from which you want to plot (be careful if your star.log
+                         output is sparse...)
+        modstop:         model to which you wish to plot
+        xlims,ylims:     plot limits, however these are somewhat obsolete now that we
+                         have modstart and modstop. Leaving them as 0. is probably
+                         no slower, and you can always zoom in afterwards in mpl
+        outfile:         'filename + extension' where you want to save the figure
+        ixaxis:          either 'log_time_left', 'age', or 'model_number'
+        xres,yres:       plot resolution. Needless to say that increasing these
+                         values will yield a nicer plot with some slow-down in
+                         plotting time. You will most commonly change xres. For a
+                         prelim plot, try xres~200, then bump it up to anywhere from
+                         1000-10000 for real nicely resolved, publication quality
+                         plots.
+        mix_zones,
+        burn_zones:      As described above, if you have more detailed output about
+                         your convection and energy generation boundaries in columns
+                         mix_type_n, mix_qtop_n, burn_type_n and burn_qtop_n, you need
+                         to specify the total number of columns for mixing zones and
+                         burning zones that you have. Can't work this out from your
+                         star.log file? Check the log_columns.list that you used, it'll
+                         be the number after "mixing regions" and "burning regions".
+                         Can't see these columns? leave it and 2 conv zones and 2 burn
+                         zones will be drawn using other data that you certainly should
+                         have in your star.log file.'''
+
+
+        xxyy=[self.get('star_age')[modstart:modstop],self.get('star_age')[modstart:modstop]]
         mup = max(float(self.get('star_mass')[0])*1.02,1.0)
-        nmodels=len(self.get('model_number'))
+        nmodels=len(self.get('model_number')[modstart:modstop])
 
         # y-axis resolution
         ny=yres
@@ -632,8 +665,8 @@ class star_log(DataPlot):
 
         engenstyle = 'full'
 
-	B1=np.random.random_sample((len(y),len(x)))
-	B2=np.random.random_sample((len(y),len(x)))
+	B1=np.zeros([len(y),len(x)],float)
+	B2=np.zeros([len(y),len(x)],float)
         try:
             self.get('burn_qtop_1')
         except:
@@ -645,12 +678,12 @@ class star_log(DataPlot):
 	        sys.stdout.flush()
 	        sys.stdout.write("\rcreating color map1 " + "...%d%%" % percent)
 	        for j in range(1,burn_zones+1):
-	            ulimit=self.get('burn_qtop_'+str(j))[i*dx]*self.get('star_mass')[i*dx]
+	            ulimit=self.get('burn_qtop_'+str(j))[modstart:modstop][i*dx]*self.get('star_mass')[modstart:modstop][i*dx]
 	            if j==1:
 	                llimit=0.0
 	            else:
-	                llimit=self.get('burn_qtop_'+str(j-1))[i*dx]*self.get('star_mass')[i*dx]
-	            btype=float(self.get('burn_type_'+str(j))[i*dx])
+	                llimit=self.get('burn_qtop_'+str(j-1))[modstart:modstop][i*dx]*self.get('star_mass')[modstart:modstop][i*dx]
+	            btype=float(self.get('burn_type_'+str(j))[modstart:modstop][i*dx])
 	            if llimit!=ulimit:
 	                for k in range(ny):
 	                    if llimit<=y[k] and ulimit>y[k] and btype>0.:
@@ -658,20 +691,20 @@ class star_log(DataPlot):
 		            if llimit<=y[k] and ulimit>y[k] and btype<0.:
 		                B2[k,i]=10**(abs(btype))
         if engenstyle == 'twozone':
-                V=np.random.random_sample((len(y),len(x)))
+                V=np.zeros([len(y),len(x)],float)
                 for i in range(len(x)):
                 # writing reading status 
                   percent = int(i*100/len(x))
                   sys.stdout.flush()
                   sys.stdout.write("\rcreating color map1 " + "...%d%%" % percent)
-                  llimitl1=self.get('epsnuc_M_1')[i*dx]/Msol
-                  ulimitl1=self.get('epsnuc_M_4')[i*dx]/Msol
-                  llimitl2=self.get('epsnuc_M_5')[i*dx]/Msol
-                  ulimitl2=self.get('epsnuc_M_8')[i*dx]/Msol
-                  llimith1=self.get('epsnuc_M_2')[i*dx]/Msol
-                  ulimith1=self.get('epsnuc_M_3')[i*dx]/Msol
-                  llimith2=self.get('epsnuc_M_6')[i*dx]/Msol
-                  ulimith2=self.get('epsnuc_M_7')[i*dx]/Msol
+                  llimitl1=self.get('epsnuc_M_1')[modstart:modstop][i*dx]/Msol
+                  ulimitl1=self.get('epsnuc_M_4')[modstart:modstop][i*dx]/Msol
+                  llimitl2=self.get('epsnuc_M_5')[modstart:modstop][i*dx]/Msol
+                  ulimitl2=self.get('epsnuc_M_8')[modstart:modstop][i*dx]/Msol
+                  llimith1=self.get('epsnuc_M_2')[modstart:modstop][i*dx]/Msol
+                  ulimith1=self.get('epsnuc_M_3')[modstart:modstop][i*dx]/Msol
+                  llimith2=self.get('epsnuc_M_6')[modstart:modstop][i*dx]/Msol
+                  ulimith2=self.get('epsnuc_M_7')[modstart:modstop][i*dx]/Msol
                   # lower thresh first, then upper thresh:
                   if llimitl1!=ulimitl1:
                       for k in range(ny):
@@ -703,12 +736,12 @@ class star_log(DataPlot):
 	          sys.stdout.flush()
 	          sys.stdout.write("\rcreating color map2 " + "...%d%%" % percent)
 	          for j in range(1,mix_zones+1):
-	            ulimit=self.get('mix_qtop_'+str(j))[i*dx]*self.get('star_mass')[i*dx]
+	            ulimit=self.get('mix_qtop_'+str(j))[modstart:modstop][i*dx]*self.get('star_mass')[modstart:modstop][i*dx]
 	            if j==1:
 	              llimit=0.0
 	            else:
-	              llimit=self.get('mix_qtop_'+str(j-1))[i*dx]*self.get('star_mass')[i*dx]
-	            mtype=self.get('mix_type_'+str(j))[i*dx]
+	              llimit=self.get('mix_qtop_'+str(j-1))[modstart:modstop][i*dx]*self.get('star_mass')[modstart:modstop][i*dx]
+	            mtype=self.get('mix_type_'+str(j))[modstart:modstop][i*dx]
 	            if llimit!=ulimit:
 	              for k in range(ny):
 		        if llimit<=y[k] and ulimit>y[k] and mtype == 1:
@@ -720,14 +753,14 @@ class star_log(DataPlot):
 	          percent = int(i*100/len(x))
 	          sys.stdout.flush()
 	          sys.stdout.write("\rcreating color map2 " + "...%d%%" % percent)
-                  ulimit=self.get('conv_mx1_top')[i*dx]*self.get('star_mass')[i*dx]
-	          llimit=self.get('conv_mx1_bot')[i*dx]*self.get('star_mass')[i*dx]
+                  ulimit=self.get('conv_mx1_top')[modstart:modstop][i*dx]*self.get('star_mass')[modstart:modstop][i*dx]
+	          llimit=self.get('conv_mx1_bot')[modstart:modstop][i*dx]*self.get('star_mass')[modstart:modstop][i*dx]
 	          if llimit!=ulimit:
 	              for k in range(ny):
 		          if llimit<=y[k] and ulimit>y[k]:
 		              Z[k,i]=1.
-                  ulimit=self.get('conv_mx2_top')[i*dx]*self.get('star_mass')[i*dx]
-	          llimit=self.get('conv_mx2_bot')[i*dx]*self.get('star_mass')[i*dx]
+                  ulimit=self.get('conv_mx2_top')[modstart:modstop][i*dx]*self.get('star_mass')[modstart:modstop][i*dx]
+	          llimit=self.get('conv_mx2_bot')[modstart:modstop][i*dx]*self.get('star_mass')[modstart:modstop][i*dx]
 	          if llimit!=ulimit:
 	              for k in range(ny):
 		          if llimit<=y[k] and ulimit>y[k]:
@@ -742,7 +775,7 @@ class star_log(DataPlot):
 	fig = pl.figure(1)
 	fig.set_size_inches(16,9)
 	fsize=20
-        ax=pl.axes()
+        ax=pl.axes([0.1,0.1,0.9,0.8])
 
 	if ixaxis == 'log_time_left':
 	# log of time left until core collapse
@@ -754,15 +787,15 @@ class star_log(DataPlot):
 	            lage[i]=np.log10(gage[-1]-gage[i]+agemin)
 	        else :
 	            lage[i]=np.log10(agemin)
-	    xxx = lage
+	    xxx = lage[modstart:modstop]
 	    print 'plot versus time left'
 	    ax.set_xlabel('$\mathrm{log}_{10}(t^*) \, \mathrm{(yr)}$',fontsize=fsize)
 	elif ixaxis =='model_number':
-	    xxx= self.get('model_number')
+	    xxx= self.get('model_number')[modstart:modstop]
 	    print 'plot versus model number'
 	    ax.set_xlabel('Model number',fontsize=fsize)
 	elif ixaxis =='age':
-	    xxx= self.get('star_age')/1.e6
+	    xxx= self.get('star_age')[modstart:modstop]/1.e6
 	    print 'plot versus age'
 	    ax.set_xlabel('Age [Myr]',fontsize=fsize)
 
@@ -778,7 +811,11 @@ class star_log(DataPlot):
             ylims[0] = 0.
             ylims[1] = mup
 
-        ax=pl.axes([xlims[0],xlims[1],ylims[0],ylims[1]])
+
+
+        print xxx[::dx]
+        print y
+        print Z
 
 	print 'plotting contours'
 	CMIX    = ax.contourf(xxx[::dx],y,Z, cmap=cmapMIX, alpha=0.5,levels=[0.5,1.5])
@@ -796,15 +833,18 @@ class star_log(DataPlot):
                 ax.contourf(xxx[::dx],y,V, cmap=cmapB1, alpha=0.5)
 
 	print 'plotting patches'
-	ax.plot(xxx[::dx],self.get('star_mass')[::dx],'k-')
+	ax.plot(xxx[::dx],self.get('star_mass')[modstart:modstop][::dx],'k-')
 
 	CBARBURN1.set_label('$|\epsilon_\mathrm{nuc}-\epsilon_{\\nu}| \; (\mathrm{erg\,g}^{-1}\mathrm{\,s}^{-1})$',fontsize=fsize)
 
 	print 'plotting abund boundaries'
-	ax.plot(xxx,self.get('h1_boundary_mass'),label='H boundary')
-	ax.plot(xxx,self.get('he4_boundary_mass'),label='He boundary')
-	ax.plot(xxx,self.get('c12_boundary_mass'),label='C boundary')
+	ax.plot(xxx,self.get('h1_boundary_mass')[modstart:modstop],label='H boundary')
+	ax.plot(xxx,self.get('he4_boundary_mass')[modstart:modstop],label='He boundary')
+	ax.plot(xxx,self.get('c12_boundary_mass')[modstart:modstop],label='C boundary')
 
+        ax.axis([xlims[0],xlims[1],ylims[0],ylims[1]])
+
+        fig.savefig(outfile)
         pl.show()
 # below are some utilities that the user typically never calls directly
